feat: Add lazy loading for book image on index.html
This commit adds lazy loading functionality for the book image on the index.html page. When the "Fetch Book Image" button is clicked, the fetchBookById() function is triggered, which makes an API call to retrieve the book image data. The image is then dynamically created and appended to the imageContainer div using the base64-encoded image data. This lazy loading implementation improves the page load performance by only loading the image when requested.