feat: Update Dockerfile and pom.xml
This commit updates the Dockerfile to use the openjdk:20 image and adds the Book-Store-App-0.0.1.jar file to the container. It also exposes port 6969 and sets the entrypoint to run the jar file.

In addition, the pom.xml file is modified to update the project version to 0.0.1 and the java.version property to 20.

These changes ensure that the application is built and deployed correctly with the updated dependencies and configurations.