Update Dockerfile and pom.xml to use Java 17 and Maven 3.8.1
This commit updates the Dockerfile to use the maven:3.8.1-openjdk-17 image for building the application. It also updates the pom.xml file to set the java.version and maven.compiler.source/target properties to 17. These changes ensure that the application is built and deployed correctly with the updated Java and Maven versions.