feat: Update dependencies and configurations for caching and lazy loading
This commit updates the dependencies in the pom.xml file to include the latest version of the springdoc-openapi-starter-webmvc-ui library (2.5.0). It also adds the necessary configurations for enabling caching and lazy loading in the application. These changes improve performance by caching the results of book-related operations and reducing the need to fetch data from the database repeatedly. Additionally, the OpenAPI documentation is enhanced with annotations for response DTOs and error responses, providing a better documentation experience for API consumers.